pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.1",
    "python-dotenv>=1.0.1",
    "twilio>=9.4.6",
]
